python/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
torch>=2.6.0
torchvision>=0.21.0
Pillow>=10.4.0
python-multipart==0.0.9
httpx==0.27.0
